Rename: 폴더명 재변경
대소문자 변경안되는 이슈
