+++ app/(route)/activity/register/_components/ActivityRegisterForm/index.tsx
@@ -8,9 +8,9 @@ import ACTIVITY_CATEGORY from '@/_constants/activity-category';
 import { addCommasToPrice, removeCommas } from '@/_utils/formatNumber';
 
 import Button from '@/_components/button';
-import Input from '@/_components/InputTest';
+import Input from '@/_components/Input';
 import SelectBox from '@/_components/SelectBox';
-import Textarea from '@/_components/TextareaTest';
+import Textarea from '@/_components/Textarea';
 
 import type { Activity } from '../../page';
 import BannerImage from '../BannerImage';
